BACKPORT: FROMLIST: MIPS: Reserve extra memory for crash dump
Traditionally, MIPS's contiguous low memory can be as less as 256M, so
crashkernel=X@Y may be unable to large enough in some cases. Moreover,
for the "multi numa node + sparse memory model" case, it is attempt to
allocate section_mem_maps on every node. Thus, if the total memory of a
node is more than 1GB, we reserve the top 128MB for the crash kernel.

Signed-off-by: Huacai Chen <chenhc@lemote.com>

Link: https://lore.kernel.org/all/1600942079-18652-2-git-send-email-chenhc@lemote.com/
[Kexy: Resolved minor conflict in arch/mips/kernel/setup.c]
Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>
Signed-off-by: WangYuli <wangyuli@uniontech.com>