drm/i915/dsb: Evade transcoder undelayed vblank when using DSB
commit 7678e089bd18b4dd61478ce728f01fd5239e97ff upstream.

We want to start the DSB execution from the transcoder's undelayed
vblank, so in order to guarantee atomicity with the all the other
mmio register writes we need to evade both vblanks.

Note that currently we don't add any vblank delay, so this is
effectively a nop. But in the future when we start to program
double buffered registers from the DSB we'll need to delay the
pipe's vblank to provide the register programming "window2"
for the DSB.

deepin-Intel-SIG: commit 7678e089bd18 drm/i915/dsb: Evade transcoder undelayed vblank when using DSB.
MTL Graphics backporting - 2024Q2

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230606191504.18099-15-ville.syrjala@linux.intel.com
Reviewed-by: Uma Shankar <uma.shankar@intel.com>
[ Quanxian Wang: amend commit log ]
Signed-off-by: Quanxian Wang <quanxian.wang@intel.com>
Signed-off-by: WangYuli <wangyuli@uniontech.com>