drm/i915: Simplify intel_cx0_program_phy_lane() with loop
commit 0f5c2e5bd2fc8d2e09043b6bb3c81a889a483997 upstream.

It is possible to generalize the "disable" value for the transmitters to
be a bit mask based on the port width and the port reversal boolean,
with a small exception for DP-alt mode with "x1" port width.

Simplify the code by using such a mask and a for-loop instead of using
switch-case statements.

v2:
  - Use (i < 2) instead of (i / 2 == 0) for PHY lane mask selection.
    (Jani)

BSpec: 64539
deepin-Intel-SIG: commit 0f5c2e5bd2fc drm/i915: Simplify intel_cx0_program_phy_lane() with loop.
MTL Graphics backporting - 2024Q2

Cc: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Gustavo Sousa <gustavo.sousa@intel.com>
Reviewed-by: Mika Kahola <mika.kahola@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230814131331.69516-3-gustavo.sousa@intel.com
[ Quanxian Wang: amend commit log ]
Signed-off-by: Quanxian Wang <quanxian.wang@intel.com>
Signed-off-by: WangYuli <wangyuli@uniontech.com>