drm/i915/vdsc: Add a check for dsc split cases
commit 611977c3e4048a7cda42a8e87dec1672f15f763b ("drm/i915/vdsc: Add a check for dsc split cases") upstream.

In intel_vdsc_get_config we only read the primary dsc engine register
and not take into account if the other dsc engine is in use and if
both registers have the same value or not this patche fixes that by
adding a check.

--v3
-Remove superfluos new line [Jani]
-Fix register naming [Jani]

--v5
-pps_temp0/pps_temp1 can be assigned where they are used [Ankit]

deepin-Intel-SIG: commit 611977c3e404 ("drm/i915/vdsc: Add a check for dsc split cases").

Signed-off-by: Suraj Kandpal <suraj.kandpal@intel.com>
Reviewed-by: Ankit Nautiyal <ankit.k.nautiyal@intel.com>
Signed-off-by: Animesh Manna <animesh.manna@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20230828054300.560559-3-suraj.kandpal@intel.com
[ Junxiao Chang: amend commit log ]
Signed-off-by: Junxiao Chang <junxiao.chang@intel.com>
Signed-off-by: WangYuli <wangyuli@uniontech.com>