wifi: rtw89: fw: move polling function of firmware path ready to an individual function
To download firmware, we need to check path is ready. There are two kinds
of path -- one is to download firmware header, and the other is to download
firmware body.

Since the polling method is different from WiFi 7 chips, make it to be
an individual function, and then we can reuse the download flow.

Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>
Signed-off-by: Kalle Valo <kvalo@kernel.org>
Link: https://lore.kernel.org/r/20230901073956.54203-2-pkshih@realtek.com
(cherry picked from commit 80e706a85cb5e1e2121e9bf09dd4098415d85142)
Signed-off-by: WangYuli <wangyuli@uniontech.com>