crypto: qat - add rp2svc sysfs attribute
commit dbc8876dd873 ("crypto: qat - add rp2svc sysfs attribute") upstream

Add the attribute `rp2svc` to the `qat` attribute group. This provides a
way for a user to query a specific ring pair for the type of service
that is currently configured for.

When read, the service will be returned for the defined ring pair.
When written to this value will be stored as the ring pair to return
the service of.

Intel-SIG: commit dbc8876dd873 ("crypto: qat - add rp2svc sysfs attribute)"
Backport to support Intel QAT in-tree driver

Signed-off-by: Ciunas Bennett <ciunas.bennett@intel.com>
Reviewed-by: Giovanni Cabiddu <giovanni.cabiddu@intel.com>
Reviewed-by: Damian Muszynski <damian.muszynski@intel.com>
Reviewed-by: Tero Kristo <tero.kristo@linux.intel.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
[ Aichun Shi: amend commit log ]
Signed-off-by: Aichun Shi <aichun.shi@intel.com>
Signed-off-by: WangYuli <wangyuli@uniontech.com>