drm/i915/lnl: Extend C10/C20 phy
commit 10184a8a7f70d28ba6aae22142a7375a8c8c1924 ("drm/i915/lnl: Extend C10/C20 phy") upstream.

For Lunar Lake, DDI-A is connected to C10 PHY, while TC1-TC3 are connected
to C20 phy, like in Meteor Lake. Update the check in intel_is_c10phy()
accordingly.

This reverts the change in commit e388ae97e225 ("drm/i915/display:
Eliminate IS_METEORLAKE checks") that turned that into a display engine
version check. The phy <-> port connection is very SoC-specific and not
related to that version.

IS_LUNARLAKE() is defined to 0 in i915 as it's expected that the
(upcoming) xe driver is the one defining the platform, with i915 only
driving the display side.

deepin-Intel-SIG: commit 10184a8a7f70 ("drm/i915/lnl: Extend C10/C20 phy").

Bspec: 70818
Reviewed-by: Gustavo Sousa <gustavo.sousa@intel.com>
Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20231026184045.1015655-2-lucas.demarchi@intel.com
[ Junxiao Chang: amend commit log ]
Signed-off-by: Junxiao Chang <junxiao.chang@intel.com>
Signed-off-by: WangYuli <wangyuli@uniontech.com>