crypto: qat - prevent underflow in rp2srv_store()
commit e53c741303a5 ("crypto: qat - prevent underflow in rp2srv_store()") upstream

The "ring" variable has an upper bounds check but nothing checks for
negatives.  This code uses kstrtouint() already and it was obviously
intended to be declared as unsigned int.  Make it so.

Intel-SIG: commit e53c741303a5 ("crypto: qat - prevent underflow in rp2srv_store())"
Backport to support Intel QAT in-tree driver

Fixes: dbc8876dd873 ("crypto: qat - add rp2svc sysfs attribute")
Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>
Acked-by: Giovanni Cabiddu <giovanni.cabiddu@intel.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
[ Aichun Shi: amend commit log ]
Signed-off-by: Aichun Shi <aichun.shi@intel.com>
Signed-off-by: WangYuli <wangyuli@uniontech.com>