AOSCOS: MIPS: lib: exclude generic memset.o if CPU_LOONGSON64 is set
In "AOSCOS: memset optimization for loongson-3", a Loongson-3-specific
memset implementation was introduced, with its own set of definitions
for `bzero' and `memset' routines. This resulted in multiple definitions
for the aforementioned assembly routines, causing vmlinux to fail to link.

Exclude generic memset.o if `CPU_LOONGSON64' is set to avoid this error.

Fixes: "AOSCOS: memset optimization for loongson-3"
Signed-off-by: Mingcong Bai <jeffbai@aosc.io>

Signed-off-by: Kexy Biscuit <kexybiscuit@aosc.io>
Signed-off-by: WangYuli <wangyuli@uniontech.com>