EDAC/amd64: Fix the calculation of instance id for Hygon family 18h model 6h
On Hygon family 18h model 6h platform, each DDR has two subchannels,
and each subchannel has a mca bank with a cs corresponded. In the
process of address translation, it needs to know the instance id
of cs to access df register correctly. The instance id of cs can be
calculated from IPID[23:20] and IPID[13]: IPID[23:20] represents
which umc controller it belongs to, and IPID[13] represents which
subchannel it is.

Fixes: 47ed942f1901 ("EDAC/amd64: Adjust UMC channel for Hygon family 18h model 6h")
Signed-off-by: Liao Xuan <liaoxuan@hygon.cn>