drm/arm: Suppress "ARM devices" menu if not essential
Manifestly, the "ARM devices" menu is appropriate only for the ARM
architecture.

Inasmuch as every single item in this menu is predicated on
(ARM || ARM64 || COMPILE_TEST), it is therefore illogical for the
"ARM devices" menu itself not to be predicated upon
(ARM || ARM64 || COMPILE_TEST), and displaying a blank menu in other
architectures is unwarranted.

Signed-off-by: WangYuli <wangyuli@uniontech.com>