MIPS: KGDB: Remove the useless "nop"
The nop instruction surrounding "breakinst:\tbreak\n\t" appears to
serve no real purpose.

Its introduction can be traced back to commit 51c6022fdb ("[PATCH]
MIPS update") within the Linux history tree [1]. This commit was
substantial, comprising 41010 lines, and provides no justification
for the insertion of this nop instruction.

Based on the MIPS architecture specification, delay slots are only
present after jump instructions or MIPS1 load instructions.
Consequently, the nop here is not intended to satisfy a delay slot
requirement.

Thus, this instruction is suspicious and should probably be removed.

[1]. https://web.git.kernel.org/pub/scm/linux/kernel/git/tglx/history.git

Cc: Maciej W. Rozycki <macro@orcam.me.uk>
Cc: Huacai Chen <chenhuacai@kernel.org>
Cc: WANG Xuerui <kernel@xen0n.name>
Cc: Felix Yan <felixonmars@archlinux.org>
Cc: Mingcong Bai <jeffbai@aosc.io>
Signed-off-by: WangYuli <wangyuli@uniontech.com>
---
NOTE:
  This patch is submitted as an RFC due to my incomplete knowledge
of the extensive history of MIPS.
  The nop instruction in question may indeed have a valid reason for
its existence, but its origins are likely too far in the past to
easily ascertain.
  I would be grateful if anyone with relevant historical information
could take the time to elaborate on the background, such as specific
models or microarchitectures that might be impacted by this change.
  Thank you all very much for your assistance.